RemotePlayer: Remove Settings writer to Files database
