+++ src/database/database-files.h
@@ -38,7 +38,14 @@ public:
 	void listPlayers(std::vector<std::string> &res);
 
 private:
-	void serialize(std::ostringstream &os, RemotePlayer *player);
+	void deSerialize(RemotePlayer *p, std::istream &is,
+		const std::string &playername, PlayerSAO *sao);
+	/*
+		serialize() writes a bunch of text that can contain
+		any characters except a '\0', and such an ending that
+		deSerialize stops reading exactly at the right point.
+	*/
+	void serialize(RemotePlayer *p, std::ostream &os);
 
 	std::string m_savedir;
 };
