+++ src/remoteplayer.cpp
@@ -83,122 +83,6 @@ RemotePlayer::RemotePlayer(const char *name, IItemDefManager *idef):
 	m_star_params = sky_defaults.getStarDefaults();
 }
 
-void RemotePlayer::serializeExtraAttributes(std::string &output)
-{
-	assert(m_sao);
-	Json::Value json_root;
-
-	const StringMap &attrs = m_sao->getMeta().getStrings();
-	for (const auto &attr : attrs) {
-		json_root[attr.first] = attr.second;
-	}
-
-	output = fastWriteJson(json_root);
-}
-
-
-void RemotePlayer::deSerialize(std::istream &is, const std::string &playername,
-		PlayerSAO *sao)
-{
-	Settings args;
-
-	if (!args.parseConfigLines(is, "PlayerArgsEnd")) {
-		throw SerializationError("PlayerArgsEnd of player " + playername + " not found!");
-	}
-
-	m_dirty = true;
-	//args.getS32("version"); // Version field value not used
-	const std::string &name = args.get("name");
-	strlcpy(m_name, name.c_str(), PLAYERNAME_SIZE);
-
-	if (sao) {
-		try {
-			sao->setHPRaw(args.getU16("hp"));
-		} catch(SettingNotFoundException &e) {
-			sao->setHPRaw(PLAYER_MAX_HP_DEFAULT);
-		}
-
-		try {
-			sao->setBasePosition(args.getV3F("position"));
-		} catch (SettingNotFoundException &e) {}
-
-		try {
-			sao->setLookPitch(args.getFloat("pitch"));
-		} catch (SettingNotFoundException &e) {}
-		try {
-			sao->setPlayerYaw(args.getFloat("yaw"));
-		} catch (SettingNotFoundException &e) {}
-
-		try {
-			sao->setBreath(args.getU16("breath"), false);
-		} catch (SettingNotFoundException &e) {}
-
-		try {
-			const std::string &extended_attributes = args.get("extended_attributes");
-			std::istringstream iss(extended_attributes);
-			Json::CharReaderBuilder builder;
-			builder.settings_["collectComments"] = false;
-			std::string errs;
-
-			Json::Value attr_root;
-			Json::parseFromStream(builder, iss, &attr_root, &errs);
-
-			const Json::Value::Members attr_list = attr_root.getMemberNames();
-			for (const auto &it : attr_list) {
-				Json::Value attr_value = attr_root[it];
-				sao->getMeta().setString(it, attr_value.asString());
-			}
-			sao->getMeta().setModified(false);
-		} catch (SettingNotFoundException &e) {}
-	}
-
-	try {
-		inventory.deSerialize(is);
-	} catch (SerializationError &e) {
-		errorstream << "Failed to deserialize player inventory. player_name="
-			<< name << " " << e.what() << std::endl;
-	}
-
-	if (!inventory.getList("craftpreview") && inventory.getList("craftresult")) {
-		// Convert players without craftpreview
-		inventory.addList("craftpreview", 1);
-
-		bool craftresult_is_preview = true;
-		if(args.exists("craftresult_is_preview"))
-			craftresult_is_preview = args.getBool("craftresult_is_preview");
-		if(craftresult_is_preview)
-		{
-			// Clear craftresult
-			inventory.getList("craftresult")->changeItem(0, ItemStack());
-		}
-	}
-}
-
-void RemotePlayer::serialize(std::ostream &os)
-{
-	// Utilize a Settings object for storing values
-	Settings args;
-	args.setS32("version", 1);
-	args.set("name", m_name);
-
-	// This should not happen
-	assert(m_sao);
-	args.setU16("hp", m_sao->getHP());
-	args.setV3F("position", m_sao->getBasePosition());
-	args.setFloat("pitch", m_sao->getLookPitch());
-	args.setFloat("yaw", m_sao->getRotation().Y);
-	args.setU16("breath", m_sao->getBreath());
-
-	std::string extended_attrs;
-	serializeExtraAttributes(extended_attrs);
-	args.set("extended_attributes", extended_attrs);
-
-	args.writeLines(os);
-
-	os<<"PlayerArgsEnd\n";
-
-	inventory.serialize(os);
-}
 
 const RemotePlayerChatResult RemotePlayer::canSendChatMessage()
 {
